+++ test/integration/src/segments/AirportListSegment.js
@@ -1,5 +1,5 @@
 import MapSegment from 'soya/lib/data/redux/segment/map/MapSegment';
-import Thunk from 'soya/lib/data/redux/Thunk.js';
+import Load from 'soya/lib/data/redux/Load';
 
 // TODO: Figure out how to do polyfill.
 // TODO: Figure out how to load client-side libraries like jQuery!
@@ -17,13 +17,9 @@ export default class AirportListSegment extends MapSegment {
     return '*';
   }
 
-  _isLoadQuery() {
-    return true;
-  }
-
-  _generateThunkFunction(thunk) {
-    var queryId = thunk.queryId;
-    thunk.func = (dispatch) => {
+  _createLoadFromQuery(query, queryId, segmentState) {
+    var load = new Load();
+    load.func = (dispatch) => {
       var result = new Promise((resolve, reject) => {
         request.get('http://localhost:8000/api/airport/list').end((err, res) => {
           if (res.ok) {
@@ -37,5 +33,6 @@ export default class AirportListSegment extends MapSegment {
       });
       return result;
     };
+    return load;
   }
 }