+++ test/integration/src/segments/BadgeSegment.js
@@ -1,5 +1,5 @@
 import MapSegment from 'soya/lib/data/redux/segment/map/MapSegment';
-import Thunk from 'soya/lib/data/redux/Thunk';
+import Load from 'soya/lib/data/redux/Load';
 
 // TODO: Figure out how to do polyfill.
 // TODO: Figure out how to load client-side libraries like jQuery!
@@ -16,13 +16,9 @@ export default class BadgeSegment extends MapSegment {
     return '*';
   }
 
-  _isLoadQuery() {
-    return true;
-  }
-
-  _generateThunkFunction(thunk) {
-    var queryId = thunk.queryId;
-    thunk.func = (dispatch) => {
+  _createLoadFromQuery(query, queryId, segmentState) {
+    var load = new Load();
+    load.func = (dispatch) => {
       var result = new Promise((resolve, reject) => {
         request.get('http://localhost:8000/api/user/badge/list').end((err, res) => {
           if (res.ok) {
@@ -36,6 +32,7 @@ export default class BadgeSegment extends MapSegment {
       });
       return result;
     };
+    return load;
   }
 
   _processRefreshRequests(segmentState, refreshRequests) {