+++ test/integration/src/segments/FibonacciTotalSegment.js
@@ -1,10 +1,6 @@
-import MapSegment from 'soya/lib/data/redux/segment/map/MapSegment.js';
-import Thunk from 'soya/lib/data/redux/Thunk.js';
+import MapSegment from 'soya/lib/data/redux/segment/map/MapSegment';
 import QueryDependencies from 'soya/lib/data/redux/QueryDependencies';
-
-// TODO: Figure out how to do polyfill.
-// TODO: Figure out how to load client-side libraries like jQuery!
-import request from 'superagent';
+import Load from 'soya/lib/data/redux/Load';
 
 import { FibonacciTotalSegmentId } from './ids.js';
 import FibonacciSegment from './FibonacciSegment.js';
@@ -22,18 +18,13 @@ export default class FibonacciTotalSegment extends MapSegment {
     return query.number;
   }
 
-  _isLoadQuery() {
-    return true;
-  }
-
-  _generateThunkFunction(thunk) {
-    var query = thunk.query;
-    var queryId = thunk.queryId;
+  _createLoadFromQuery(query, queryId, segmentState) {
+    var load = new Load();
     var dependencies = QueryDependencies.serial(Promise);
 
     dependencies.add('total', FibonacciSegment.id(), query);
-    thunk.dependencies = dependencies;
-    thunk.func = (dispatch) => {
+    load.dependencies = dependencies;
+    load.func = (dispatch) => {
       var i, total = 0, n, resultArray = dependencies.getResult('total').data.split(' ');
       for (i = 0; i < resultArray.length; i++) {
         n = resultArray[i];
@@ -43,5 +34,6 @@ export default class FibonacciTotalSegment extends MapSegment {
       var actionObj = this._createSyncLoadActionObject(queryId, total);
       return dispatch(actionObj);
     };
+    return load;
   }
 }