+++ test/integration/src/segments/LifetimeSessionSegment.js
@@ -1,6 +1,6 @@
-import MapSegment from 'soya/lib/data/redux/segment/map/MapSegment.js';
-import Thunk from 'soya/lib/data/redux/Thunk.js';
-import Cookie from 'soya/lib/http/Cookie.js';
+import MapSegment from 'soya/lib/data/redux/segment/map/MapSegment';
+import Cookie from 'soya/lib/http/Cookie';
+import Load from 'soya/lib/data/redux/Load';
 
 // TODO: Figure out how to do polyfill.
 // TODO: Figure out how to load client-side libraries like jQuery!
@@ -20,17 +20,13 @@ export default class LifetimeSessionSegment extends MapSegment {
     return 'default';
   }
 
-  _isLoadQuery() {
-    return true;
-  }
-
-  _generateThunkFunction(thunk) {
-    var queryId = thunk.queryId;
+  _createLoadFromQuery(query, queryId, segmentState) {
+    var load = new Load();
     var lifetimeCookie = this._cookieJar.read(LIFETIME_COOKIE_NAME);
     var sessionCookie = this._cookieJar.read(SESSION_COOKIE_NAME);
     if (lifetimeCookie != null && sessionCookie != null) {
       // Just re-use what we already have in cookie.
-      thunk.func = (dispatch) => {
+      load.func = (dispatch) => {
         dispatch(this._createSyncLoadActionObject(queryId, {
           lifetime: lifetimeCookie,
           session: sessionCookie
@@ -40,7 +36,7 @@ export default class LifetimeSessionSegment extends MapSegment {
       return;
     }
 
-    thunk.func = (dispatch) => {
+    load.func = (dispatch) => {
       var result = new Promise((resolve, reject) => {
         request.get('http://localhost:8000/api/context').end((err, res) => {
           if (res.ok) {
@@ -58,5 +54,6 @@ export default class LifetimeSessionSegment extends MapSegment {
       });
       return result;
     };
+    return load;
   }
 }